Return instance of log in callback rather than only errors
It makes more sense to return the instance of Log, for a given lint run, in place
of the copy of just the errors, for the lint run, in the callback. This remove
the need to expose it as an instance method of the linter.

Closes #18

update npm scripts and documentation